Moved ``string[]`` returning methods, add ZenExpansions.
